refactor: extract discardHand and drawHand to utils
