+++ src/components/Toolbar.jsx
@@ -1,4 +1,5 @@
-import { ModalShowing, NORMAL_HAND_SIZE } from "../data/constants";
+import { ModalShowing } from "../data/constants";
+import { discardHand, drawNewHand } from "../utils/HandOperations";
 import TurnTracker from "./TurnTracker";
 
 
@@ -27,48 +28,6 @@ export default function Toolbar({ setModalShowing, cardsInHand, setCardsInHand, 
             {/* TODO: add more tools here */}
         </div>
     );
-}
-
-
-
-/**
- * Draws {NORMAL_HAND_SIZE} new cards if the users hand is empty. Otherwise does nothing.
- */
-function drawNewHand(cardsInHand, setCardsInHand, cardsInDeck, setCardsInDeck, cardsInDiscard, setCardsInDiscard) {
-    if (cardsInHand.length > 0)
-        return;
-
-    const deckSize = cardsInDeck.length;
-
-    //Draw cards from the deck
-    let startIndex = Math.max(0, deckSize - NORMAL_HAND_SIZE);
-    let endIndex = deckSize;
-    let cardsDrawn = cardsInDeck.slice(startIndex, endIndex);
-
-    if (cardsDrawn.length < NORMAL_HAND_SIZE) {
-        //Draw remaining cards from discard pile, and reset the deck.
-        const additionalCardsNeeded = Math.min(NORMAL_HAND_SIZE - cardsDrawn.length, cardsInDiscard.length);
-        let cardsDrawnFromDiscard = cardsInDiscard.slice(0, additionalCardsNeeded);
-        let cardsLeftInDiscard = cardsInDiscard.slice(additionalCardsNeeded, cardsInDiscard.length);
-
-        cardsDrawn = [...cardsDrawn, ...cardsDrawnFromDiscard];
-        setCardsInDeck(cardsLeftInDiscard);
-        setCardsInDiscard([]);
-    }
-    else {
-        //Remove drawn cards from deck
-        setCardsInDeck(prev => prev.slice(0, startIndex));
-    }
-
-
-    setCardsInHand(cardsDrawn.reverse());
-}
-
-
-
-function discardHand(cardsInHand, setCardsInHand, setCardsInDiscard) {
-    setCardsInDiscard(prev => [...prev, ...cardsInHand.reverse()]);
-    setCardsInHand([]);
 }
